+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.29
+
+* `ComparableEnum` will now correctly disassemble other enums